requirements.txt
flask
torch
transformers
opencv-python
numpy
Pillow
pyttsx3